Update from JenkinsRsWlan5:release/v1.0_sync_20250314

+++ component/os/os_wrapper/include/os_wrapper_time.h
@@ -40,7 +40,7 @@ uint32_t rtos_time_get_current_system_time_ms(void);
  * @note   This interface does not consider systick overflow issues.
  * @retval
  */
-uint32_t rtos_time_get_current_system_time_us(void);
+uint64_t rtos_time_get_current_system_time_us(void);
 
 /**
  * @brief  Rreturn value to in nanoseconds.
